+++ jwp-basic-gradle/src/main/java/core/web/mvcframework/adapter/HandlerAdapter.java
@@ -1,6 +1,6 @@
-package core.mvcframework.adapter;
+package core.web.mvcframework.adapter;
 
-import core.mvcframework.ModelAndView;
+import core.web.mvcframework.ModelAndView;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
