+++ jwp-basic-gradle/src/main/java/core/web/mvcframework/mapping/annotation/AnnotationHandlerMapping.java
@@ -1,13 +1,12 @@
-package core.mvcframework.mapping.annotation;
+package core.web.mvcframework.mapping.annotation;
 
 import com.google.common.collect.Maps;
 import core.annotation.Controller;
 import core.annotation.RequestMapping;
 import core.annotation.RequestMethod;
 import core.di.ApplicationContext;
-import core.mvcframework.mapping.HandlerMapping;
+import core.web.mvcframework.mapping.HandlerMapping;
 import java.lang.reflect.Method;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
