+++ jwp-basic-gradle/src/main/java/next/config/MyWebApplicationInitializer.java
@@ -1,8 +1,8 @@
 package next.config;
 
-import core.mvcframework.DispatcherServlet;
-import core.mvcframework.mapping.HandlerMapping;
-import core.mvcframework.mapping.annotation.AnnotationHandlerMapping;
+import core.web.mvcframework.DispatcherServlet;
+import core.web.mvcframework.mapping.HandlerMapping;
+import core.web.mvcframework.mapping.annotation.AnnotationHandlerMapping;
 import core.web.WebApplicationInitializer;
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
@@ -18,10 +18,12 @@ public class MyWebApplicationInitializer implements WebApplicationInitializer {
     public void onStartup(final ServletContext servletContext) throws ServletException {
         HandlerMapping handlerMapping = new AnnotationHandlerMapping("next", "core");
         handlerMapping.initialize();
-        ServletRegistration.Dynamic dispatcher = servletContext.addServlet("dispatcher", new DispatcherServlet(handlerMapping));
+        ServletRegistration.Dynamic dispatcher = servletContext.addServlet("dispatcher",
+                new DispatcherServlet(handlerMapping));
         dispatcher.setLoadOnStartup(1);
         dispatcher.addMapping("/");
 
         log.info("Start MyWebApplication Initializer");
     }
+
 }
