+++ jwp-basic-gradle/src/main/java/next/controller/qna/QuestionController.java
@@ -4,8 +4,8 @@ import core.annotation.Controller;
 import core.annotation.Inject;
 import core.annotation.RequestMapping;
 import core.annotation.RequestMethod;
-import core.mvcframework.ModelAndView;
-import core.mvcframework.controller.AbstractAnnotationController;
+import core.web.mvcframework.ModelAndView;
+import core.web.mvcframework.controller.AbstractAnnotationController;
 import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
