+++ jwp-basic-gradle/src/main/java/next/controller/user/legacy/UserListController.java
@@ -1,7 +1,7 @@
 package next.controller.user.legacy;
 
-import core.mvcframework.ModelAndView;
-import core.mvcframework.controller.AbstractController;
+import core.web.mvcframework.ModelAndView;
+import core.web.mvcframework.controller.AbstractController;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
