+++ jwp-basic-gradle/src/test/java/core/nmvc/AnnotationHandlerMappingTest.java
@@ -2,9 +2,9 @@ package core.nmvc;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import core.mvcframework.ModelAndView;
-import core.mvcframework.mapping.annotation.AnnotationHandlerMapping;
-import core.mvcframework.mapping.annotation.HandlerExecution;
+import core.web.mvcframework.ModelAndView;
+import core.web.mvcframework.mapping.annotation.AnnotationHandlerMapping;
+import core.web.mvcframework.mapping.annotation.HandlerExecution;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
