+++ jwp-basic-gradle/src/test/java/core/nmvc/MyController.java
@@ -3,8 +3,8 @@ package core.nmvc;
 import core.annotation.Controller;
 import core.annotation.RequestMapping;
 import core.annotation.RequestMethod;
-import core.mvcframework.ModelAndView;
-import core.mvcframework.view.JspView;
+import core.web.mvcframework.ModelAndView;
+import core.web.mvcframework.view.JspView;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
